mc6

+++ pFREYA_tester/prova_comparator4.py
@@ -51,7 +51,7 @@ data.append(float(config.lecroy.query('C1:CRVA? HREL').split(',')[2])) #C1 è il
 
 while(True):
     time.sleep(1)
-    val = (float)(config.lecroy.query("VBS? 'app.Measure.P5.value'"))
+    val = config.lecroy.query("VBS? 'app.Measure.P5.value'")
     print(str(val))
